Moving model tests into per-backend test suite
